Ghost Cat AI (In Progress)
Ghost Cat AI for the memory game. I will walk to the piano, sit on the seat and when the player is close enough it will play a randomised tune. The ghost cat will then play a note and the player needs to copy the ghost cat until they have completed the whole sequence.

Missing some features still.